Fix any wrong sanitization and correct highscores

+++ Final_Project/include/highscore_manager.h
@@ -7,8 +7,8 @@
 namespace battle_ship {
 class highscore_manager {
 public:
-  static std::vector<std::tuple<std::string, int>> all_highscores;
-  static void add_highscore(std::tuple<std::string, int> h);
+  static std::vector<std::tuple<std::string, size_t>> all_highscores;
+  static void add_highscore(std::tuple<std::string, size_t> h);
   static void initialise_highscores();
 };
 } // namespace battle_ship
